wellness/DeepSeekBackend/requirements.txt
flask==2.3.3
ollama==0.1.6
python-dotenv==1.0.0
requests==2.31.0 